backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
paddlepaddle==2.5.2
paddleocr==2.7.3
tesseract==0.1.3
pytesseract==3.3.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
pix2tex[api]==0.1.2
doctr[tf]==0.7.0
mathpix-markdown-it==1.2.0
pandoc==2.3
python-docx==1.1.0
lxml==4.9.3
requests==2.31.0
httpx==0.25.2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
loguru==0.7.2
prometheus-client==0.19.0